chore: rename authentication fields -> replace clients into provider
