+++ code/api/server/src/main/java/com/decathlon/ara/configuration/authentication/provider/custom/AuthenticationCustomValueConfiguration.java
@@ -15,7 +15,7 @@
  *                                                                            *
  ******************************************************************************/
 
-package com.decathlon.ara.configuration.authentication.clients.custom;
+package com.decathlon.ara.configuration.authentication.provider.custom;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
