+++ code/api/server/src/main/java/com/decathlon/ara/configuration/authentication/provider/custom/token/AuthenticationCustomTokenConfiguration.java
@@ -15,22 +15,17 @@
  *                                                                            *
  ******************************************************************************/
 
-package com.decathlon.ara.configuration.authentication.clients.custom.token;
+package com.decathlon.ara.configuration.authentication.provider.custom.token;
 
+import com.decathlon.ara.configuration.authentication.provider.custom.AuthenticationCustomValueConfiguration;
 import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
 
 @Data
-public class AuthenticationCustomTokenFieldsConfiguration {
+@Configuration
+@ConfigurationProperties("authentication.provider.custom.token")
+public class AuthenticationCustomTokenConfiguration extends AuthenticationCustomValueConfiguration {
 
-    private String id;
-
-    private String access;
-
-    private String refresh;
-
-    private String expiration;
-
-    private String type;
-
-    private String scope;
+    private AuthenticationCustomTokenFieldsConfiguration fields;
 }
