+++ code/api/server/src/main/java/com/decathlon/ara/configuration/authentication/provider/github/AuthenticationGithubConfiguration.java
@@ -15,7 +15,7 @@
  *                                                                            *
  ******************************************************************************/
 
-package com.decathlon.ara.configuration.authentication.clients.github;
+package com.decathlon.ara.configuration.authentication.provider.github;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -23,7 +23,7 @@ import org.springframework.context.annotation.Configuration;
 
 @Data
 @Configuration
-@ConfigurationProperties("authentication.clients.github")
+@ConfigurationProperties("authentication.provider.github")
 public class AuthenticationGithubConfiguration {
 
     private Boolean enabled;
