+++ code/api/server/src/main/resources/application-auth.properties.example
@@ -26,51 +26,51 @@
 authentication.enabled=false
 
 # Github
-authentication.clients.github.enabled=true
-authentication.clients.github.client-id=your_github_client_id
-authentication.clients.github.client-secret=your_github_client_secret
+authentication.provider.github.enabled=true
+authentication.provider.github.client-id=your_github_client_id
+authentication.provider.github.client-secret=your_github_client_secret
 
 # Google
-authentication.clients.google.enabled=true
-authentication.clients.google.client-id=your_google_client_id
-authentication.clients.google.client-secret=your_google_client_secret
+authentication.provider.google.enabled=true
+authentication.provider.google.client-id=your_google_client_id
+authentication.provider.google.client-secret=your_google_client_secret
 
 # Custom
-authentication.clients.custom.enabled=true
-authentication.clients.custom.displayed-name=Company Name
-authentication.clients.custom.login-uri=the_custom_token_uri
+authentication.provider.custom.enabled=true
+authentication.provider.custom.displayed-name=Company Name
+authentication.provider.custom.login-uri=the_custom_token_uri
 
-authentication.clients.custom.token.uri=your_custom_token_uri
-authentication.clients.custom.token.method=POST
-authentication.clients.custom.token.header-values=your_token_header_attribute1,your_token_header_value1|your_token_header_attribute2,your_token_header_value2
-authentication.clients.custom.token.body-values=your_token_body_attribute1,your_token_body_value1|your_token_body_attribute2,your_token_body_value2|your_code_attribute,{{code}}
+authentication.provider.custom.token.uri=your_custom_token_uri
+authentication.provider.custom.token.method=POST
+authentication.provider.custom.token.header-values=your_token_header_attribute1,your_token_header_value1|your_token_header_attribute2,your_token_header_value2
+authentication.provider.custom.token.body-values=your_token_body_attribute1,your_token_body_value1|your_token_body_attribute2,your_token_body_value2|your_code_attribute,{{code}}
 
-authentication.clients.custom.token.fields.id=your_token_id_field
-authentication.clients.custom.token.fields.access=your_token_access_field
-authentication.clients.custom.token.fields.refresh=your_token_refresh_field
-authentication.clients.custom.token.fields.expiration=your_token_expiration_field
-authentication.clients.custom.token.fields.type=your_token_type_field
-authentication.clients.custom.token.fields.scope=your_token_scope_field
+authentication.provider.custom.token.fields.id=your_token_id_field
+authentication.provider.custom.token.fields.access=your_token_access_field
+authentication.provider.custom.token.fields.refresh=your_token_refresh_field
+authentication.provider.custom.token.fields.expiration=your_token_expiration_field
+authentication.provider.custom.token.fields.type=your_token_type_field
+authentication.provider.custom.token.fields.scope=your_token_scope_field
 
-authentication.clients.custom.user.uri=your_custom_user_uri
-authentication.clients.custom.user.method=POST
-authentication.clients.custom.user.header-values=a_user_header_attribute,a_user_header_value|a_user_header_attribute_with_variables,{{token_type}} {{token_value}}
-authentication.clients.custom.user.body-values=your_custom_user_body_attribute1,your_custom_user_body_value1|your_custom_user_body_attribute2,your_custom_user_body_value2
+authentication.provider.custom.user.uri=your_custom_user_uri
+authentication.provider.custom.user.method=POST
+authentication.provider.custom.user.header-values=a_user_header_attribute,a_user_header_value|a_user_header_attribute_with_variables,{{token_type}} {{token_value}}
+authentication.provider.custom.user.body-values=your_custom_user_body_attribute1,your_custom_user_body_value1|your_custom_user_body_attribute2,your_custom_user_body_value2
 
-authentication.clients.custom.user.fields.id=your_user_id_field
-authentication.clients.custom.user.fields.name=your_user_name_field
-authentication.clients.custom.user.fields.login=your_user_login_field
-authentication.clients.custom.user.fields.email=your_user_email_field
-authentication.clients.custom.user.fields.picture-url=your_user_picture-url_field
+authentication.provider.custom.user.fields.id=your_user_id_field
+authentication.provider.custom.user.fields.name=your_user_name_field
+authentication.provider.custom.user.fields.login=your_user_login_field
+authentication.provider.custom.user.fields.email=your_user_email_field
+authentication.provider.custom.user.fields.picture-url=your_user_picture-url_field
 
-authentication.clients.custom.token-validation.uri=your_custom_token_validation_uri
-authentication.clients.custom.token-validation.method=POST
-authentication.clients.custom.token-validation.header-values=your_token-validation_header_attribute1,your_token-validation_header_value1 {{token_value}}|your_token-validation_header_attribute2,your_token-validation_header_value2
-authentication.clients.custom.token-validation.body-values=authorization_attribute,{{token_value}}|your_token-validation_body_attribute1,your_token-validation_body_value1|your_token-validation_body_attribute2,your_token-validation_body_value2
-authentication.clients.custom.token-validation.validation-field.name=the_token_validation_field
-authentication.clients.custom.token-validation.validation-field.expected-value=the_token_validation_expected_value
-authentication.clients.custom.token-validation.validation-field.remaining-time=expire-in
-authentication.clients.custom.token-validation.validation-field.expiration-timestamp=expiration-date
+authentication.provider.custom.token-validation.uri=your_custom_token_validation_uri
+authentication.provider.custom.token-validation.method=POST
+authentication.provider.custom.token-validation.header-values=your_token-validation_header_attribute1,your_token-validation_header_value1 {{token_value}}|your_token-validation_header_attribute2,your_token-validation_header_value2
+authentication.provider.custom.token-validation.body-values=authorization_attribute,{{token_value}}|your_token-validation_body_attribute1,your_token-validation_body_value1|your_token-validation_body_attribute2,your_token-validation_body_value2
+authentication.provider.custom.token-validation.validation-field.name=the_token_validation_field
+authentication.provider.custom.token-validation.validation-field.expected-value=the_token_validation_expected_value
+authentication.provider.custom.token-validation.validation-field.remaining-time=expire-in
+authentication.provider.custom.token-validation.validation-field.expiration-timestamp=expiration-date
 
 # JWT
 authentication.jwt-token.using-https=false
