+++ code/api/server/src/test/java/com/decathlon/ara/configuration/authentication/provider/custom/AuthenticationCustomValueConfigurationTest.java
@@ -15,7 +15,7 @@
  *                                                                            *
  ******************************************************************************/
 
-package com.decathlon.ara.configuration.authentication.clients.custom;
+package com.decathlon.ara.configuration.authentication.provider.custom;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
